Began working on clientside prediction
didn't implement anything yet but created a plan for how it should work in order to make the player feel as though they have no latency with their movements. once this is done we can work on fun stuff like inventory system, item picking up and dropping, the base and some enemy ai eventually. this is all distant stuff however