Allowed items to be spawned on the server and picked up by clients
next job is to allow other clients to know who has what items and display their items with a worldmodel. Also some slight folder organisation to help with adding items in the future both on the server and clients. regardless, still like 4 seperate files to add so in future i need to make that more streamlined with shared libraries or something similar which will take ages to set up