viewmodels for other client's items
one error that clients which just join the server don't see peoples items until they update them.
need to add shooting and health packets and run bullets on the server with a client predicted bullet on client or not depending on latency.
hitmarkers and various audio elements needed before playtest